pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "az-search-to-milvus"
version = "0.1.0"
description = "Migration tool from Azure AI Search to Milvus/Zilliz — schema conversion, data migration, and validation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "az-ai-search-to-milvus contributors" },
]
keywords = ["azure", "ai-search", "milvus", "zilliz", "vector-database", "migration"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "azure-search-documents>=11.6.0",
    "azure-identity>=1.17.0",
    "pymilvus>=2.5.0,<3.0.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
    "mypy>=1.10",
]

[project.scripts]
az-search-to-milvus = "az_search_to_milvus.cli:main"

[project.urls]
Homepage = "https://github.com/Beginnersguide138/az-ai-search-to-milvus"
Repository = "https://github.com/Beginnersguide138/az-ai-search-to-milvus"

[tool.hatch.build.targets.wheel]
packages = ["src/az_search_to_milvus"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
